+++ packages/restapi/src/lib/payloads/constants.ts
@@ -1,5 +1,5 @@
 export interface ChainIdToSourceType {
-  [key: number]: string
+  [key: number]: string;
 }
 
 export const CHAIN_ID_TO_SOURCE: ChainIdToSourceType = {
@@ -26,19 +26,19 @@ export const SOURCE_TYPES = {
   OPTIMISM_TESTNET: 'OPTIMISM_TESTNET',
   POLYGON_ZK_EVM_TESTNET: 'POLYGON_ZK_EVM_TESTNET',
   POLYGON_ZK_EVM_MAINNET:'POLYGON_ZK_EVM_MAINNET',
-  THE_GRAPH: 'THE_GRAPH'
+  THE_GRAPH: 'THE_GRAPH',
+  PUSH_VIDEO: 'PUSH_VIDEO',
 };
 
-export enum IDENTITY_TYPE  {
+export enum IDENTITY_TYPE {
   MINIMAL = 0,
   IPFS = 1,
   DIRECT_PAYLOAD = 2,
-  SUBGRAPH = 3
-};
+  SUBGRAPH = 3,
+}
 
-export enum NOTIFICATION_TYPE  {
+export enum NOTIFICATION_TYPE {
   BROADCAST = 1,
   TARGETTED = 3,
-  SUBSET = 4
-};
-
+  SUBSET = 4,
+}