+++ packages/restapi/src/lib/user/index.ts
@@ -2,5 +2,10 @@ export * from './createUser';
 export * from './getFeeds';
 export * from './getSubscriptions';
 export * from './getUser';
-export * from "./getDelegations";
+export * from './getDelegations';
 export * from './getUsersBatch';
+export * from './upgradeUser';
+import { update } from './auth.updateUser';
+export const auth = {
+  update,
+};